# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HR AI platform API with multilingual profile parsing powered by Gemini",
        "title": "HR AI API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/profile-fields/{id}/texts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile-field-texts"
                ],
                "summary": "List all text translations for a profile field",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile field ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/http.ProfileFieldTextResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile-field-texts"
                ],
                "summary": "Create a text translation for a profile field",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile field ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Text data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateProfileFieldTextRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/http.ProfileFieldTextResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profile-fields/{id}/texts/{lang}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile-field-texts"
                ],
                "summary": "Get a profile field text by field ID and language",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile field ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language code (uz, ru, en)",
                        "name": "lang",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.ProfileFieldTextResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile-field-texts"
                ],
                "summary": "Update a profile field text",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile field ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language code (uz, ru, en)",
                        "name": "lang",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated text data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateProfileFieldTextRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.ProfileFieldTextResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "profile-field-texts"
                ],
                "summary": "Delete a profile field text",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile field ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language code (uz, ru, en)",
                        "name": "lang",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/profile/parse": {
            "post": {
                "description": "Accepts JSON with user_input text, or multipart/form-data with either user_input field or a file (PDF, PNG, JPG, TXT). Parses the profile and stores multilingual translations (uz, ru, en).",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile-parse"
                ],
                "summary": "Parse a user profile from text or file using Gemini AI",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Text input (for JSON content type)",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/http.ProfileParseTextRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Text input (for multipart)",
                        "name": "user_input",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "File upload: PDF, PNG, JPG, TXT (max 20MB)",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.ProfileParseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "http.CreateProfileFieldTextRequest": {
            "type": "object",
            "required": [
                "content",
                "lang"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "is_source": {
                    "type": "boolean"
                },
                "lang": {
                    "type": "string"
                },
                "model_version": {
                    "type": "string"
                }
            }
        },
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "http.ProfileFieldTextResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "is_source": {
                    "type": "boolean"
                },
                "lang": {
                    "type": "string"
                },
                "model_version": {
                    "type": "string"
                },
                "profile_field_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "http.ProfileParseEducationResponse": {
            "type": "object",
            "properties": {
                "degree": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "field_of_study": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "institution": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "texts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.ProfileParseItemTextResponse"
                    }
                }
            }
        },
        "http.ProfileParseExperienceResponse": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "projects": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "texts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.ProfileParseItemTextResponse"
                    }
                },
                "web_site": {
                    "type": "string"
                }
            }
        },
        "http.ProfileParseFieldResponse": {
            "type": "object",
            "properties": {
                "field_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "texts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.ProfileParseFieldTextResponse"
                    }
                }
            }
        },
        "http.ProfileParseFieldTextResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "is_source": {
                    "type": "boolean"
                },
                "lang": {
                    "type": "string"
                },
                "model_version": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "http.ProfileParseItemTextResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "is_source": {
                    "type": "boolean"
                },
                "lang": {
                    "type": "string"
                },
                "model_version": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "http.ProfileParseResponse": {
            "type": "object",
            "properties": {
                "education": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.ProfileParseEducationResponse"
                    }
                },
                "experience": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.ProfileParseExperienceResponse"
                    }
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.ProfileParseFieldResponse"
                    }
                },
                "source_lang": {
                    "type": "string"
                }
            }
        },
        "http.ProfileParseTextRequest": {
            "type": "object",
            "required": [
                "user_input"
            ],
            "properties": {
                "user_input": {
                    "type": "string"
                }
            }
        },
...
        "http.UpdateProfileFieldTextRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "model_version": {
                    "type": "string"
                }
            }
        },
...
    }
}